Issue #20747: move Input files to compiled folder
