+++ config/google-java-format/excluded/compilable-input-paths.txt
@@ -22,6 +22,7 @@ src/it/resources/com/google/checkstyle/test/chapter4formatting/rule413emptyblock
 src/it/resources/com/google/checkstyle/test/chapter4formatting/rule42blockindentation/InputClassWithChainedMethods3.java
 src/it/resources/com/google/checkstyle/test/chapter4formatting/rule42blockindentation/InputIndentationCodeBlocks.java
 src/it/resources/com/google/checkstyle/test/chapter4formatting/rule43onestatement/InputOneStatementPerLine.java
+src/it/resources/com/google/checkstyle/test/chapter4formatting/rule43onestatement/InputOneStatementPerLine2.java
 src/it/resources/com/google/checkstyle/test/chapter4formatting/rule44columnlimit/InputColumnLimit.java
 src/it/resources/com/google/checkstyle/test/chapter4formatting/rule44columnlimit/InputTextBlockColumnLimit.java
 src/it/resources/com/google/checkstyle/test/chapter4formatting/rule44columnlimit/InputColumnLimitEdgeCase.java
