+++ config/jsoref-spellchecker/whitelist.words
@@ -682,7 +682,6 @@ Jax
 jb
 JBoss
 JButton
-JCheck
 JCombo
 JComponent
 jcp
