+++ src/it/java/com/google/checkstyle/test/chapter4formatting/rule43onestatement/OneStatementPerLineTest.java
@@ -42,12 +42,12 @@ public class OneStatementPerLineTest extends AbstractGoogleModuleTestSupport {
 
     @Test
     public void testOneStatementInput() throws Exception {
-        verifyWithWholeConfig(getNonCompilablePath("InputOneStatementPerLine2.java"));
+        verifyWithWholeConfig(getPath("InputOneStatementPerLine2.java"));
     }
 
     @Test
     public void testOneStatementInputFormatted() throws Exception {
-        verifyWithWholeConfig(getNonCompilablePath("InputFormattedOneStatementPerLine2.java"));
+        verifyWithWholeConfig(getPath("InputFormattedOneStatementPerLine2.java"));
     }
 
 }
